Refactored the drag and drop logic into multiple smaller functions
